+++ src/OFS/interfaces.py
@@ -682,7 +682,6 @@ class IObjectManager(IZopeObject, ICopyContainer, INavigation, IManageable,
     manage_main = Attribute(""" """)
     manage_index_main = Attribute(""" """)
     manage_addProduct = Attribute(""" """)
-    manage_importExportForm = Attribute(""" """)
 
     def all_meta_types(interfaces=None):
         """
@@ -759,17 +758,6 @@ class IObjectManager(IZopeObject, ICopyContainer, INavigation, IManageable,
 
     def tpValues():
         """Return a list of subobjects, used by tree tag.
-        """
-
-    def manage_exportObject(id='', download=None, toxml=None,
-                            RESPONSE=None, REQUEST=None):
-        """Exports an object to a file and returns that file."""
-
-    def manage_importObject(file, REQUEST=None, set_owner=1):
-        """Import an object from a file"""
-
-    def _importObjectFromFile(filepath, verify=1, set_owner=1):
-        """
         """